# docs/api/dcrs-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DCRS API for the Mitra mobile app (the Audit Assistant)",
    "version": "1.3.0",
    "summary": "Read and change DCRS data as a signed-in person, through DCRS's own rules.",
...
  },
  "servers": [
    {
      "url": "http://{host}:{port}",
      "description": "The DCRS server on the factory network (API_PORT, 4000 unless set).",
      "variables": {
        "host": {
          "default": "localhost",
          "description": "The DCRS computer's name or address."
        },
        "port": {
          "default": "4000"
        }
      }
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/escalations": {
      "get": {
        "tags": [
          "Mitra",
          "Read"
        ],
        "operationId": "escalations",
        "summary": "READ — the super admin's escalations (the super admin only)",
        "description": "The escalations DCRS's server raised for the super admin (backend/escalation.ts, REQUIREMENTS §75): somebody with 3 or more late submissions, or 2 or more records never done, in the last 30 days — named by name, or as the department where several share the work — what the bell in DCRS shows. `summary` is the line Mitra adds to her live facts for the super admin in DCRS, in the same words. Each escalation says who, the counts in a sentence, the documents with the most misses and up to ten of the records behind them. They name people, so they are the super admin's alone: anybody else is refused 403 super-admin-only. Reading them changes nothing; acknowledging one stays in DCRS.",
        "parameters": [
          {
            "name": "open",
            "in": "query",
            "required": false,
            "description": "1 (when left out): the ones not yet acknowledged. 0: every one raised or grown in the last 30 days.",
            "schema": {
              "type": "string",
              "enum": [
                "1",
                "0",
                "true",
                "false"
              ],
              "default": "1"
            }
          },
          {
            "$ref": "#/components/parameters/ClientName"
          }
        ],
        "responses": {
          "200": {
            "description": "The escalations.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Escalations"
                },
                "example": {
                  "open": true,
                  "today": "2026-10-02",
                  "week": "2026-W40",
                  "rule": {
                    "late": 3,
                    "neverDone": 2,
                    "windowDays": 30
                  },
                  "summary": "Escalated to the super admin, not yet acknowledged: Kapila Barad (3 late).",
                  "waiting": 1,
                  "total": 1,
                  "escalations": [
                    {
                      "id": "11",
                      "kind": "person",
                      "subjectName": "Kapila Barad",
                      "department": "QC",
                      "departmentName": "Quality Control",
                      "period": "2026-W40",
                      "late": 3,
                      "neverDone": 0,
                      "sentence": "3 late in the 30 days to 02-Oct-2026 — F-QC-30: 3 late",
                      "raisedAt": "2026-10-01T04:30:00.000Z",
                      "updatedAt": "2026-10-02T04:30:00.000Z",
                      "acknowledged": false,
                      "acknowledgedBy": null,
                      "acknowledgedAt": null,
                      "window": {
                        "from": "2026-09-03",
                        "to": "2026-10-02"
                      },
                      "worst": [
                        {
                          "documentId": "qc-viscosity",
                          "what": "F-QC-30",
                          "late": 3,
                          "neverDone": 0
                        }
                      ],
                      "records": [
                        {
                          "id": "rec-munyt2ti-14-azkgv0",
                          "documentId": "qc-viscosity",
                          "what": "F-QC-30",
                          "dueDate": "2026-09-28",
                          "outcome": "late",
                          "daysLate": 2
                        }
                      ]
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/NotSignedIn"
          },
          "403": {
            "description": "Not the super admin (super-admin-only). Also password-change-required, as on every /api/v1 route.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "super-admin-only": {
                    "value": {
                      "error": "Escalations are the super admin's alone: they name people, so no other account is shown them.",
                      "code": "super-admin-only"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "The escalations could not be read from the database (database-unavailable). Try again in a moment.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "database-unavailable": {
                    "value": {
                      "error": "DCRS could not read the escalations just now. Try again in a moment.",
                      "code": "database-unavailable"
                    }
                  }
                }
              }
            }
          }
        },
        "x-dcrs-kind": "read",
        "x-mitra-answer": "Mitra's live facts for the super admin in DCRS: engine/assistantLocal.ts buildAssistantContext (REQUIREMENTS §75)"
      }
    },
...
    "/api/jobs/run": {
      "post": {
        "tags": [
          "Change"
        ],
        "operationId": "jobsRun",
        "summary": "CHANGE - run one of the server's jobs now (the super admin only)",
        "description": "The super admin's \"run it now\" (REQUIREMENTS §97, the Notifications page's \"The server's jobs\"): morning-prepare prepares the day's records on the server, the known parts only, never a reading; notify works out each person's notifications and pushes their phones. Both are safe to run again (a second prepare finds nothing to prepare, and nobody is pushed twice for the same thing). For the plant's today it is that period's run, and the schedule does not repeat it. Each run is a line of the activity log. The test servers run with JOBS=0 and call this instead.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Run.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRunResponse"
                }
              }
            }
          },
          "400": {
            "description": "A job, a day or a time this route does not accept.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "bad-request": {
                    "value": {
                      "error": "job must be one of: escalation, weekly-digest, morning-prepare, notify.",
                      "code": "bad-request"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NotSignedIn"
          },
          "403": {
            "description": "Anybody but the super admin.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "super-admin-only": {
                    "value": {
                      "error": "Only the super admin runs the server's jobs.",
                      "code": "super-admin-only"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "The dcrs_session cookie's value from POST /api/auth/login. Valid until the close of the day it was started (the sign-in answer's session.endsAt)."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dcrs_session"
      }
    },
    "parameters": {
      "ClientName": {
        "name": "X-Client-Name",
        "in": "header",
        "required": false,
        "description": "The calling server's name, written in DCRS's record history and activity log (\"Through Mitra mobile app: ...\"). Printable characters, at most 40; a colon is dropped. Absent: \"DCRS API\".",
        "schema": {
          "type": "string",
          "maxLength": 40
        },
        "example": "Mitra mobile app"
      },
...
    },
    "responses": {
      "NotSignedIn": {
        "description": "No session, a session whose day has closed (a session ends at the close of the day it was started: the end of the working hours, or midnight for the super admin), or an account switched off. Sign in again.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "not-signed-in": {
                "value": {
                  "error": "Not signed in, or the session has ended. Sign in again with POST /api/auth/login.",
                  "code": "not-signed-in"
                }
              }
            }
          }
        }
      },
...
      "BadRequest": {
        "description": "A query value is not one this route accepts.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "bad-request": {
                "value": {
                  "error": "status must be \"open\", \"closed\" or \"all\".",
                  "code": "bad-request"
                }
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Every refusal. `error` is a sentence to show a person; `code` is for the program. A refusal outside the staff's working hours also says when they start again (`opensAt`).",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "opensAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Only with the code `outside-working-hours`: the next moment DCRS opens (the start of the next working day), or null when the plant's calendar has no working day in the year ahead."
          }
        }
      },
...
      "Escalation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "person",
              "department"
            ]
          },
          "subjectName": {
            "type": "string"
          },
          "department": {
            "type": [
              "string",
              "null"
            ]
          },
          "departmentName": {
            "type": [
              "string",
              "null"
            ]
          },
          "period": {
            "type": "string",
            "description": "The ISO week it belongs to, such as 2026-W40."
          },
          "late": {
            "type": "integer"
          },
          "neverDone": {
            "type": "integer"
          },
          "sentence": {
            "type": "string"
          },
          "raisedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "acknowledged": {
            "type": "boolean"
          },
          "acknowledgedBy": {
            "type": [
              "string",
              "null"
            ]
          },
          "acknowledgedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "window": {
            "type": "object",
            "properties": {
              "from": {
                "type": "string"
              },
              "to": {
                "type": "string"
              }
            }
          },
          "people": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "A department's accounts, by name."
          },
          "worst": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "documentId": {
                  "type": "string"
                },
                "what": {
                  "type": "string"
                },
                "late": {
                  "type": "integer"
                },
                "neverDone": {
                  "type": "integer"
                }
              }
            }
          },
          "records": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "documentId": {
                  "type": "string"
                },
                "what": {
                  "type": "string"
                },
                "dueDate": {
                  "type": "string"
                },
                "outcome": {
                  "type": "string",
                  "enum": [
                    "late",
                    "never done"
                  ]
                },
                "daysLate": {
                  "type": "integer"
                }
              }
            }
          }
        }
      },
      "Escalations": {
        "type": "object",
        "properties": {
          "open": {
            "type": "boolean"
          },
          "today": {
            "type": "string"
          },
          "week": {
            "type": "string"
          },
          "rule": {
            "type": "object",
            "properties": {
              "late": {
                "type": "integer"
              },
              "neverDone": {
                "type": "integer"
              },
              "windowDays": {
                "type": "integer"
              }
            }
          },
          "summary": {
            "type": "string",
            "description": "The line Mitra adds for the super admin in DCRS, or that nothing is waiting."
          },
          "waiting": {
            "type": "integer",
            "description": "How many are not yet acknowledged."
          },
          "total": {
            "type": "integer"
          },
          "escalations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Escalation"
            }
          }
        }
      },
...
      "JobRunRequest": {
        "type": "object",
        "required": [
          "job"
        ],
        "properties": {
          "job": {
            "type": "string",
            "enum": [
              "escalation",
              "weekly-digest",
              "morning-prepare",
              "notify"
            ],
            "description": "morning-prepare: the day's records prepared on the server (the known parts only). notify: each person's notifications worked out and their phones pushed. escalation and weekly-digest: the super admin's (REQUIREMENTS §75)."
          },
          "today": {
            "type": "string",
            "format": "date",
            "description": "A day of the caller's own (a test's); left out, the plant's today."
          },
          "time": {
            "type": "string",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "description": "The moment, plant time, the notification jobs are worked out at; left out, now."
          }
        }
      },
      "JobRunResponse": {
        "type": "object",
        "properties": {
          "job": {
            "type": "string"
          },
          "today": {
            "type": "string",
            "format": "date"
          },
          "outcome": {
            "type": "string",
            "description": "How it went, in a line (the activity log has the same)."
          },
          "result": {
            "type": "object",
            "description": "The job's own figures. morning-prepare: date, prepared, sheetsMade, modules [{module, count}], records. notify: date, time, planned, written, resolved, escalations, cleaned, push."
          },
          "claimed": {
            "type": [
              "string",
              "null"
            ],
            "description": "The period this run stood for (the schedule then leaves it alone), or null for a day of the caller's own or one already run."
          }
        }
      }
    }
  }
}